Assign Random Character Index
Since the grid has already been filled with the character of the chosen
index, n times, based on the number of times provided as parameter to
the shuffleBoard function, the rest of the grid should be filled-up with
characters from random indeces of array _characters. The goal of the
added code is to find these random indeces with the condition: index is
not equal to the index you want to find AND not equal to zero, which
implies the NONE/null/empty character.